Collapse Room into Rooms; move sidebar components into Layout
